[FEAT, REFACTOR] 홈, 포스트, 친구, 매칭 일부 변경 및 추가

+++ src/main/java/com/backend/together/domain/post/service/PostServiceImpl.java
@@ -1,5 +1,12 @@
 package com.backend.together.domain.post.service;
 
+import com.backend.together.domain.block.Entity.Block;
+import com.backend.together.domain.block.service.BlockServiceImpl;
+import com.backend.together.domain.member.entity.MemberEntity;
+import com.backend.together.domain.member.repository.MemberRepository;
+import com.backend.together.domain.post.dto.PostRequestDTO;
+import com.backend.together.global.apiPayload.code.status.ErrorStatus;
+import com.backend.together.global.apiPayload.exception.handler.CustomHandler;
 import com.backend.together.global.enums.Category;
 import com.backend.together.domain.post.repository.HashtagRepository;
 import com.backend.together.global.enums.Gender;
@@ -8,8 +15,14 @@ import com.backend.together.domain.post.Post;
 import com.backend.together.domain.post.repository.PostRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,7 +30,11 @@ import java.util.List;
 import java.util.Optional;
 @Slf4j
 @Service
+@RequiredArgsConstructor
+@Transactional
 public class PostServiceImpl implements PostService {
+    private final MemberRepository memberRepository;
+    private final BlockServiceImpl blockService;
 
     @Autowired
     private PostRepository repository;
@@ -28,24 +45,15 @@ public class PostServiceImpl implements PostService {
     @PersistenceContext
     private EntityManager entityManager;
 
-    public String testService() {
-        Post post = Post.builder().title("first post").build();
-        repository.save(post);
-        Post savedPost = repository.findById(post.getId()).get();
-        return savedPost.getTitle();
-    }
-
     @Override
-    public void createPost(Post post) {
-        //validate
+    public Post createPost(PostRequestDTO requestDTO, Long memberId) {
+        MemberEntity member = memberRepository.findByMemberId(memberId)
+                .orElseThrow(() -> new CustomHandler(ErrorStatus.MEMBER_NOT_FOUND));
 
+        Post newPost = PostRequestDTO.toEntity(requestDTO, member);
+        repository.save(newPost);
 
-        repository.save(post);
-
-
-
-
-
+        return newPost;
     }
 //    public void saveHashtag(Question question, List<String> tagNames) {
 //
@@ -59,8 +67,18 @@ public class PostServiceImpl implements PostService {
 //    }
 
     @Override
-    public void deletePost(Long postId) {
-        //validate
+    public void deletePost(Long postId, Long memberId) {
+        MemberEntity member = memberRepository.findByMemberId(memberId)
+                        .orElseThrow(() -> new CustomHandler(ErrorStatus.MEMBER_NOT_FOUND));
+
+        Post post = repository.findById(postId)
+                        .orElseThrow(() -> new CustomHandler(ErrorStatus.POST_NOT_FOUND));
+
+        // 글쓴이가 아니면 삭제 불가 by. Jayde
+        if (!post.getMember().equals(member)) {
+            throw new CustomHandler(ErrorStatus.INVALID_APPROACH);
+        }
+
         repository.deleteById(postId);
     }
 
@@ -75,8 +93,10 @@ public class PostServiceImpl implements PostService {
     }
 
     public List<Post> retrievePostByMemberId(Long id){
-        System.out.println(repository.findAll());
-        return repository.findPostByMemberId(id);
+        MemberEntity member = memberRepository.findByMemberId(id)
+                .orElseThrow(() -> new CustomHandler(ErrorStatus.MEMBER_NOT_FOUND));
+
+        return repository.findPostByMember(member);
     }
 
     @Override
@@ -92,6 +112,50 @@ public class PostServiceImpl implements PostService {
         return postByTitleContainingOrContentContaining;
     }
 
+    @Override
+    public void updateView(Post post) {
+        Long postView = post.getView();
+
+        post.updateView(postView+1);
+    }
+
+    @Override
+    public Page<Post> getPostByCategorySort(String category, String sort, Integer page) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Long userId = Long.parseLong(authentication.getName());
+
+        Category postCategory = switch (category.toUpperCase()) {
+            case "HOBBY" -> Category.HOBBY;
+            case "PLAY" -> Category.PLAY;
+            case "EAT" -> Category.EAT;
+            case "EXERCISE" -> Category.EXERCISE;
+            default -> throw new CustomHandler(ErrorStatus.CATEGORY_NOT_FOUND);
+        };
+
+        List<MemberEntity> blockedList = blockService.getBlockedMember(userId).stream().map(Block::getBlocked).toList();
+        Sort sortBy = getSort(sort);
+        if (blockedList.isEmpty()) {
+            return repository.findAllByCategory(postCategory, PageRequest.of(page, 12, sortBy));
+        } else {
+            return repository.findAllByCategoryAndMemberNotIn(postCategory, blockedList, PageRequest.of(page, 12, sortBy));
+        }
+    }
+
+    @Override
+    public Page<Post> getPostByMemberId(Long memberId, Integer page) {
+        MemberEntity member = memberRepository.findByMemberId(memberId)
+                .orElseThrow(() -> new CustomHandler(ErrorStatus.MEMBER_NOT_FOUND));
+
+        return repository.findAllByMember(member, PageRequest.of(page, 4, Sort.by("createdAt").descending()));
+    }
+
+    private Sort getSort(String sortBy) {
+        return switch (sortBy) {
+            case "popularity" -> Sort.by("view").descending();
+            case "recent" -> Sort.by("createdAt").descending();
+            default -> throw new CustomHandler(ErrorStatus.UNSUPORRTED_SORT);
+        };
+    }
 
     public List<Post> retrievePostsByCategory(Category category) {
         List<Post> postByCategory = repository.findPostByCategory(category);
@@ -107,21 +171,13 @@ public class PostServiceImpl implements PostService {
     }
 
 
-    /* Views Counting */
-    /* Views Counting */
-    @Override
-    @Transactional
-    public void updateView(Long id) {
-        repository.updateView(id);
-    }
-
-
     private static void validate(Post post) {
         if(post == null){
             log.warn("Entity cannot be null.");
         }
 
-        if(post.getMemberId() == null){
+        assert post != null;
+        if(post.getMember().getMemberId() == null){
             log.warn("Unknown user.");
             throw new RuntimeException("Unknown user.");
         }
